backend/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
arrow==0.13.1
asgiref==3.4.1
astor==0.8.1
astroid==2.5
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
audioread==2.1.9
auto-face-recognition==0.0.3
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==19.10b0
bleach==3.3.0
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.1
clipboard==0.0.4
cloudpickle==1.6.0
colorama==0.4.4
comtypes==1.1.10
constantly==15.1.0
cookiecutter==1.7.2
coverage==5.5
cryptography==3.4.7
cvzone==1.1.1
cycler==0.10.0
Cython==0.29.23
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
django-unused-media==0.2.2
djangorestframework==3.13.1
dlib==19.22.0
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
flake8==3.9.0
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
gast==0.3.3
gcloud==0.18.3
geojson==2.5.0
google-api-client==3.14.159265359.post1
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
googlesearch-python==1.0.1
googletrans==3.0.0
grpcio==1.36.1
gTTS==2.2.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hiredis==2.0.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageai==2.1.6
imageio==2.9.0
imageio-ffmpeg==0.4.4
imagesize==1.2.0
importlib-metadata==3.10.0
imutils==0.5.4
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jaraco.context==4.0.0
JarvisAI==3.1
jedi==0.17.2
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==0.17.0
Js2Py==0.71
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jws==0.1.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
keyring==23.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lazyme==0.0.23
librosa==0.8.0
llvmlite==0.36.0
lxml==4.6.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.2
mccabe==0.6.1
mediapipe==0.8.6.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
more-itertools==8.8.0
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
mysql-connector==2.2.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
notebook==6.4.0
numba==0.53.1
numpy==1.21.1
numpydoc==1.1.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.1.1
pipwin==0.5.0
playsound==1.2.2
pluggy==0.13.1
pooch==1.4.0
poyo==0.5.0
proglog==0.1.9
prometheus-client==0.10.1
prompt-toolkit==3.0.17
protobuf==3.14.0
psutil==5.8.0
psycopg2==2.9.2
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.11.0
pydocstyle==6.1.1
pyflakes==2.2.0
pygame==2.1.0
PyGetWindow==0.0.9
Pygments==2.9.0
pyjokes==0.6.0
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.7.4
pyls-black==0.4.6
pyls-spyder==0.3.2
pymongo==3.12.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyowm==3.2.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
pyreadline==2.1
Pyrebase4==4.5.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.27
pyserial==3.5
pySmartDL==1.3.4
PySocks==1.7.1
pytest==6.2.4
pytest-warnings==0.3.1
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-jwt==2.0.1
python-language-server==0.36.2
python-slugify==5.0.2
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==5.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.2
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.5.0
roman==3.3
rope==0.18.0
rsa==4.7.2
Rtree==0.9.7
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.4.1
screen-brightness-control==0.9.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==21.1.0
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.15.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
soupsieve==2.2.1
SpeechRecognition==3.8.1
Sphinx==4.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==5.0.0
spyder-kernels==2.0.1
sqlparse==0.4.1
sweetify==2.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.4
testcode==1.0.0
testpath==0.4.4
text-unidecode==1.3
textdistance==4.2.1
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
three-merge==0.1.1
tinycss==0.4
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.50.2
traitlets==5.0.5
transformers==4.3.2
twilio==6.60.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.4
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.36.2
whichcraft==0.6.1
whitenoise==5.3.0
wikipedia==1.4.0
Wikipedia-API==0.5.4
win-inet-pton==1.1.0
wincertstore==0.2
WMI==1.5.1
wolframalpha==5.0.0
wrapt==1.12.1
xmltodict==0.12.0
yapf==0.31.0
yarl==1.6.3
youtube-dl==2021.6.6
zipp==3.4.1
zope.interface==5.4.0
